fix portable object hit detection
Ignore alpha values when doing hit detection on portable objects. Also
clears inventories on enable to make sure the player's inventory isn't
remembered between sessions. I don't expect to disable an inventory at
any point as those are part of the game state, but if this changes I'll
need to reevaluate this approach.