localization
Hooked up localization and the new dialog system.

Refactored dismiss callback to be on cues instead of on the dialog box.
Now the sender can specify the callback easily and we could have one on
each cue if we wanted - e.g. we could change a sprite mid dialog.

I think this is enough to get going with the rest of the game. No more
fiddling with dialog / localization until I actually need a change.