fix scaling issues
Fiddled with dialog a bit more because I couldn't help it.

Switched the root canvas to use "Scale with Screen Size" and fixed the
shop to work with this.

Set the aspect ratio to be stuck at 3:2.

Switched image compression to high quality (normal looks like garbage).

Updated portable objects to get the scale factor from the root canvas
during dragging.