added tree
Added a tree that still needs art and some layout work, but the
inventory aspect of it is at least working as expected. I ended up
shuffling some stuff around to make this work better and I think this
has a better separation of concerns and is easier to setup in the
inspector.