cashout
Added cashout functionality to the suburb. Now when the player leaves
they're paid relative to the work they did. Still need art assets and to
do a first pass at the suburb entropy system.

This introduces a modal dialog to confirm that the player wants to leave
as they're not allowed back in the suburb after they quit for the day.