random polygon layout controller
Adds a random polygon layout controller that supports disjoint polygons
via an array for positioning objects. This is currently used by the
trees and the lawns in the suburb to find random positions that 1) are
within the bounds of the polygon(s) and 2) do not overlap any other
components based on a minimum spacing radius.