video store and localization changes
Started working on the video store and refactored the localization
engine.

Using enums was becoming troublesome as I needed to update too many
places and needed a single massive enum to contain all the strings. The
built in UI for displaying enums was awful for long lists as well.

I've switched to a dictionary with string keys and values which means
defining the string in the dict and in the inspector. Compile time
checks are broken, so this kind of sucks too.

Ideally I'd have a custom editor control that makes sure I don't use any
missing keys, but I'd rather work on the game than the editor for now.
Tech debt for later?