inspector close support
Adds the ability to close the inspector by clicking on the screen.

Fixes the LocalizedText issue of not being able to set the text prior to
awake. Also added first-party support to the GetText function to
special-case empty and null strings as an empty string.

Added the dragging canvas and inspector to the root canvas.